requirements.txt
Flask
pika
SQLAlchemy
psycopg2-binary
requests
python-dotenv
minio
pycountry
